pyproject.toml
[project]
name = "juturna"
version = "0.1.0"
description = "Juturna core library"
authors = [
    { name = "Antonio Bevilacqua", email = "b3by.in.th3.sky@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

[tool.poetry.dependencies]
ffmpeg-python = ">=0.2.0,<0.3.0"
opencv-python = "4.11.0.86,<5.0.0.0"
av = "^14.2.0"
numpy = "2.2.3,<3.0.0"
requests = "2.32.3,<3.0.0"
websockets = "^15.0.1"

# http wrapper dependencies
fastapi = { version = ">=0.115.12,<0.116.0", optional = true, extras = ["httpwrapper"] }
fastapi-cli = { version = ">=0.0.7,<0.0.8", optional = true, extras = ["httpwrapper"] }

# dev dependencies
pytest = { version = "^8.3.5", optional = true, extras = ["dev"] }
sphinx = { version = "^8.2.3", optional = true, extras = ["dev"] }
sphinxcontrib-apidoc = { version = "^0.5.0", optional = true, extras = ["dev"] }
sphinxcontrib-napoleon = { version = "^0.7", optional = true, extras = ["dev"] }
sphinx-autoapi = { version = "^3.6.0", optional = true, extras = ["dev"] }
sphinx-autobuild = { version = "^2024.10.3", optional = true, extras = ["dev"] }
sphinx-rtd-theme = { version = "^3.0.2", optional = true, extras = ["dev"] }
pydata-sphinx-theme = { version = "^0.16.1", optional = true, extras = ["dev"] }
sphinx-design = { version = "^0.6.1", optional = true, extras = ["dev"] }
sphinx-new-tab-link = {version = "^0.8.0", optional = true, extras = ["dev"] }

[tool.poetry]
package-mode = true
packages = [
    { include = "juturna" }
]

[tool.poetry.extras]
httpwrapper = ["fastapi", "fastapi-cli"]
dev = ["pytest",
       "sphinx",
       "sphinx-autoapi",
       "sphinx-autobuild",
       "sphinxcontrib-apidoc",
       "sphinxcontrib-napoleon",
       "sphinx-rtd-theme",
       "pydata-sphinx-theme",
       "sphinx-design",
       "sphinx_new_tab_link"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
sphinx = "^8.2.3"
sphinxcontrib-apidoc = "^0.5.0"
sphinxcontrib-napoleon = "^0.7"
sphinx-autoapi = "^3.6.0"
sphinx-autobuild = "^2024.10.3"
sphinx-rtd-theme = "^3.0.2"
pydata-sphinx-theme = "^0.16.1"
sphinx-design = "^0.6.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
